refactor contextual menu header function
